+++ src/VanillaPermission/Permission.php
@@ -134,82 +134,16 @@ class Permission
      */
     public function getOnly(array $ruleNames)
     {
-        // Store pre-allowed rules.
-        // Basically it'll do a simple check if rule exists on array.
-        // Eg. "users.create" will be valid, even if "users" was not defined.
-        // It'll be resolved in next step.
-        $allowedRules = $this->filterPreAllowedRules($this->sortLevel($ruleNames));
-
-        // Store post-allowed rules and your names.
-        // After allow some rule, it'll marked as post-allowed.
-        // So it make easy allow sub-rules of parent.
-        $postAllowedRules      = [ ];
-        $postAllowedRulesNames = [ ];
-
-        // Next step will unset all rules that not have defined parents.
-        foreach ($allowedRules as $allowedRule) {
-            // All zero-level rule is truly allowed.
-            // Eg. "users", "billings", ...
-            if (strpos($allowedRule->name, '.') === false) {
-                $postAllowedRules[]      = $allowedRule;
-                $postAllowedRulesNames[] = $allowedRule->name;
-                continue;
-            }
-
-            // Else, check if the parent of this rule was post-allowed.
-            $allowedRuleParent = substr($allowedRule->name, 0, strrpos($allowedRule->name, '.'));
-            if (in_array($allowedRuleParent, $postAllowedRulesNames, true)) {
-                $postAllowedRules[]      = $allowedRule;
-                $postAllowedRulesNames[] = $allowedRule->name;
-                continue;
-            }
-        }
-
         // Returns a new permission rules.
-        $permission        = new self;
-        $permission->rules = array_values($postAllowedRules);
-
-        return $permission;
-    }
-
-    /**
-     * Get allowed rules without check parents.
-     *
-     * @param  string[] $ruleNames Rule names.
-     *
-     * @return PermissionRule[]
-     */
-    private function filterPreAllowedRules(array $ruleNames)
-    {
-        $allowedRules = [ ];
+        $permission = new self;
 
         foreach ($this->getAll() as $rule) {
             if (in_array($rule->name, $ruleNames, true)) {
-                $allowedRules[] = $rule;
+                $permission->rules[] = $rule;
             }
         }
 
-        return $allowedRules;
-    }
-
-    /**
-     * Sort a rules array by level.
-     *
-     * @param  string[] $ruleNames Rule names.
-     *
-     * @return string[]
-     */
-    private function sortLevel(array $ruleNames)
-    {
-        $levelCount = [ ];
-
-        foreach ($ruleNames as $ruleName) {
-            $levelCount[$ruleName] = substr_count($ruleName, '.');
-        }
-
-        asort($levelCount);
-
-        return array_keys($levelCount);
+        return $permission;
     }
 
     /**